earth_world_model/requirements-prep.txt
numpy>=1.26
pandas>=2.2
PyYAML>=6.0
xarray>=2024.6.0
zarr>=2.18.0
numcodecs>=0.13.0
fsspec>=2024.0.0
huggingface_hub>=0.28.0
pyarrow>=15.0.0
google-cloud-storage>=3.0.0
rasterio>=1.3.0
